requirements.txt
discord.py==2.6.0
dpython-dotenv==1.0.0
